chore: update module structure for docs
