+++ src/error.rs
@@ -1,3 +1,4 @@
+//! Types for working with errors produced by es-entity.
 use thiserror::Error;
 
 #[derive(Error, Debug)]
