+++ src/one_time_executor.rs
@@ -1,3 +1,6 @@
+//! Type-safe wrapper to ensure one database operation per executor
+use crate::operation::AtomicOperation;
+
 pub struct OneTimeExecutor<'c, E>
 where
     E: sqlx::PgExecutor<'c>,
@@ -39,5 +42,41 @@ where
         A: 'q + Send + sqlx::IntoArguments<'q, sqlx::Postgres>,
     {
         query.fetch_optional(self.executor).await
+    }
+}
+
+pub trait IntoOneTimeExecutor<'c>: IntoOneTimeExecutorAt<'c> + 'c {}
+impl<'c, T> IntoOneTimeExecutor<'c> for T where T: IntoOneTimeExecutorAt<'c> + 'c {}
+
+pub trait IntoOneTimeExecutorAt<'c> {
+    type Executor: sqlx::PgExecutor<'c>;
+
+    fn into_executor(self) -> OneTimeExecutor<'c, Self::Executor>
+    where
+        Self: 'c;
+}
+
+impl<'c> IntoOneTimeExecutorAt<'c> for &sqlx::PgPool {
+    type Executor = &'c sqlx::PgPool;
+
+    fn into_executor(self) -> OneTimeExecutor<'c, Self::Executor>
+    where
+        Self: 'c,
+    {
+        OneTimeExecutor::new(self)
+    }
+}
+
+impl<'c, O> IntoOneTimeExecutorAt<'c> for &mut O
+where
+    O: AtomicOperation,
+{
+    type Executor = &'c mut sqlx::PgConnection;
+
+    fn into_executor(self) -> OneTimeExecutor<'c, Self::Executor>
+    where
+        Self: 'c,
+    {
+        OneTimeExecutor::new(self.as_executor())
     }
 }
